updated how Zoom opens
Instead of having the Zoom application open then waiting for a human input to start the meeting, I will have python open the zoom link with the meeting id and password found in the .env
This skips human input and in my opinion is better that what i had. Only downside is the application will have to work with a meeting that uses the same meeting code and password.